+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This changelog also contains important changes in dependencies.
 - (usvg) SVG Path length calculation that affects `startOffset` property in `textPath`.
 - (usvg) Fix `feImage` resolving when the linked element has
   `opacity`, `clip-path`, `mask` and/or `filter` attributes.
+- (usvg) Fix chained `feImage` resolving.
 - CLI arguments processing.
 
 ## [0.14.0] - 2021-03-06